Update: 簡易版のヘッダ、フッタナビを追加
