+++ app/pages/_app.tsx
@@ -1,6 +1,8 @@
 import App, { Container } from 'next/app';
 import Router from 'next/router';
-import Link from 'next/link';
+import GlobalHeader from '../components/header';
+import GlobalNavi from '../components/navigation';
+import styles from '../style/layout/page.module.scss';
 import '../style/style.scss';
 
 export default class MyApp extends App {
@@ -18,34 +20,11 @@ export default class MyApp extends App {
     const { Component, pageProps } = this.props;
 
     return (
-      <>
-        <header>ヘッダー</header>
+      <div className={`${styles['l-page']}`}>
+        <GlobalHeader />
         <Component {...pageProps} />
-        <nav>
-          <ul>
-            <li>
-              <Link href="/">
-                <a className="u-link">index</a>
-              </Link>
-            </li>
-            <li>
-              <Link href="/analytics">
-                <a className="u-link">analytics</a>
-              </Link>
-            </li>
-            <li>
-              <Link href="/research">
-                <a className="u-link">research</a>
-              </Link>
-            </li>
-            <li>
-              <Link href="/profile">
-                <a className="u-link">profile</a>
-              </Link>
-            </li>
-          </ul>
-        </nav>
-      </>
+        <GlobalNavi />
+      </div>
     );
   }
 }
