+++ app/pages/profile.tsx
@@ -2,9 +2,9 @@ import styles from '../style/module/profile.module.scss';
 
 function Research() {
   return (
-    <div className={`${styles['p-profile']} l-main`}>
+    <main className={`${styles['p-profile']} l-main`}>
       <p>自分の情報</p>
-    </div>
+    </main>
   );
 }
