+++ app/pages/ranking.tsx
@@ -1,13 +1,13 @@
-import styles from '../style/module/analytics.module.scss';
+import styles from '../style/module/ranking.module.scss';
 import { format } from 'date-fns';
 import { ja } from 'date-fns/locale';
 
 function Research() {
   return (
-    <div className={`${styles['p-analytics']} l-main`}>
+    <main className={`${styles['p-ranking']} l-main`}>
       <p>{format(new Date(), 'MMMdo(EEEEE) a', { locale: ja })}</p>
-      <p>チャートや損益</p>
-    </div>
+      <p>ここでみんなのカブ価がみられる</p>
+    </main>
   );
 }
