+++ app/pages/trading.tsx
@@ -1,13 +1,13 @@
-import styles from '../style/module/research.module.scss';
+import styles from '../style/module/trading.module.scss';
 import { format } from 'date-fns';
 import { ja } from 'date-fns/locale';
 
 function Research() {
   return (
-    <div className={`${styles['p-research']} l-main`}>
+    <main className={`${styles.trading} l-main`}>
       <p>{format(new Date(), 'MMMdo(EEEEE) a', { locale: ja })}</p>
-      <p>ここでみんなのカブ価がみられる</p>
-    </div>
+      <p>収支と、売買ログ</p>
+    </main>
   );
 }
