+++ app/style/module/trading.module.scss
@@ -1,4 +1,3 @@
 @import '../_variable';
-
-.p-analytics {
+.trading {
 }
