+++ app/style/style.scss
@@ -2,4 +2,5 @@
 @import '_variable';
 
 body {
+  background-color: $c_darkness;
 }
